First steps towards transitive scalar replacement
So far, in a program like:

    my $x = 0e0;
    my $y = 4e2;
    for ^10_000_000 {
        my $a = $y.sin;
        $x = $x + $a.abs
    }

Then we would be able to do scalar replacement on the Scalar `$a` and
the result of the `.abs` method. However, we could not do it for the
result of `$y.sin` because it was considered to escape at the point it
was bound into the `$!value` attribute of the Scalar `$a`.

This change lets us do that, so that we now avoid allocations for the
the result Num of `$y.sin` too. That leaves the only repeated allocation
in the program being the boxing of the result `$x + $a.abs`, which will
escape due to it being assigned into `$x` (and analyzing loops is still
a good amount of work away, and would need integration with OSR in order
to help.)

It is the minimal thing that can help, however there's still a number of
problems: it doesn't yet deopt properly in such cases, nor does it
properly track escape dependencies, so if the outer thing escapes then
the inner thing must be considered to also.