fix problems with profiling; dogbert17++
it used to be the case that the result datastructure was created
after the GC had run (for synchronization purposes), but before
anything unmarked in the gen2 was freed. This meant that the
datastructure created in the profiler never had the "seen" bits
set and were thus immediately considered "okay to be freed".

Also, the collected profile data was being added to the gc worklist
after it was evaluated and immediately before it was freed, thus
effectively preventing it from being marked (even though I don't
think it's actually ever possible for it to become part of a GC
mark phase, but this is more correct anyway)