Leave all big integer op devirt to the EA
Previously, we had:

1. A set of devirtualized ops for a few common big integer operations.
   We rewrote big integer opts into these in the post-inline pass.
2. EA over these ops.

However, the EA can do a better job of step 1 too, since it knows how to
deal with the incoming operands and result types separately. That means
it could potentially deal with `1 + (2 but Foo)`, which we could not
have done before due to inconsistent operand types. This also gets rid
of the need for the intermediate devirtualized ops, which all also dealt
in checking if we can use the integer cache; now we'll be able to handle
that in one place, if we actually have to box into a real object.

Overall, this comes out simpler, and more scalable for handling more of
the big integer ops.