[JIT] Live ranges are given register set
Decode the register spec into a register set during live range
determination. Merging register requirements is then an
bitset-union-operation. Next step is choosing a live range to spill
based on the requesting live range permissible set (to ensure that we
release a floating point register for a floating point value, and a
GPR register for an integer/pointer value).